libnetwork: re-enabe TestExternalKeyWithReexec test
Support for "libnetwork-setkey" re-exec was added in libnetwork in commit
[6175353964906b1279cd3477437b15cca9b1a2e0][1] which included a test for
both Sandbox.SetKey, and setting the key through the "libnetwork-setkey"
re-exec.

During the integration with [containerd 1.0][2], it was found that the re-exec hook
used the wrong type:

> In https://github.com/docker/libnetwork/blob/df3f8a10796e2173732aeafa41af45f970ced9f6/sandbox_externalkey_unix.go#L48
> the code is expecting libcontainer.State as input but data provided by
> libcontainer actually has the type libcontainer/configs.HookState.
>
> That means that this data doesn't contain netns path and the SetKey should not work.

This issue was fixed in [e84aad941e0577ee501de9b5ba9667e2f5daf38f][3], but
lacking a test-environment, the `TestExternalKeyWithReexec` test was removed.

[1]: https://github.com/moby/libnetwork/commit/6175353964906b1279cd3477437b15cca9b1a2e0
[2]: https://github.com/moby/moby/pull/20662
[3]: https://github.com/moby/libnetwork/commit/e84aad941e0577ee501de9b5ba9667e2f5daf38f

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>