libnetwork/iptables: make newConnection() slightly more idiomatic
Don't consider a connection successful if firewalld is not running. If
it's not running, return a suitable error message. We only use the D-Bus
connection to communicate with firewalld, so no need to return a connection
if it's not running.

Also slightly reduce depending on the global "firewalld" variable, by passing
the connection to use to checkRunning.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>