libnetwork/iptables: move setting up signals into handleSignals
newConnection was setting up these listeners, but in the event we
failed to check if firewalld was running, we would error-out, and
close the connection.

Looking at b052827e025267336f0d426df44ec536745821f8, all of this code
is directly related to handling signals, so let's combine this into the
handleSignals method.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>