pkg/plugins: pluginrpc-gen: fix go generate for volume driver
The volume driver plugin generated code was manually updated in
b15f8d2d4f054a87052a7065c50441f7e8479fa9, so generating the code
would no longer produce the same output.

This patch updates the generator code to allow regenerating the volume
driver plugin.

- update buildImports to allow additional imports to be passed
- add a conditional block to the template for volumeDriver
- add a conditional block to set the right timeout, depending on what
  method we're generating code for.
- slightly adjust the template to make the output format match the
  existing code.

Some of this is a bit quirky, so we should look for a more generic option,
but I tried to preserve the ability to generate this code, so that it
can evolve if things change.

To verify the generated code:

    GO111MODULE=off go install ./pkg/plugins/pluginrpc-gen
    GO111MODULE=off go generate ./volume/drivers

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>