libnetwork: implement Endpoint.Statistics()
This is a horrible kludge, but not /the worst/ as a reference to the
underlying `osl.Sandbox` is part of the `Sandbox`; so keeping a
reference to the underlying `osl.Interface` in the `Endpoint` is about
the same level of impure.

The lifecycle and assignment here is bad, but that can be fixed during
future refactors, and keeps this diff minimal.

Because we don't use these abstractions on Windows, it is buildtagged to
Linux.

Signed-off-by: Bjorn Neergaard <bjorn.neergaard@docker.com>