libnetwork: Endpoint: return early if no agent was found
This removes redundant nil-checks in Endpoint.deleteServiceInfoFromCluster
and Endpoint.addServiceInfoToCluster.

These functions return early if the network is not ["cluster eligible"][1],
and the function used for that (`Network.isClusterEligible`) requires the
[agent to not be `nil`][2].

This check moved around a few times ([3][3], [4][4]), but was originally
added in [libnetwork 1570][5] which, among others, tried to avoid a nil-pointer
exception reported in [moby 28712][6], which accessed the `Controller.agent`
[without locking][7]. That issue was addressed by adding locks, adding a
`Controller.getAgent` accessor, and updating deleteServiceInfoFromCluster
to use a local var. It also sprinkled this `nil` check to be on the safe
side, but as `Network.isClusterEligible` already checks for the agent
to not be `nil`, this should not be redundant.

[1]: https://github.com/moby/moby/blob/5b53ddfcdd1c0721f875d3e237078ab7de891a57/libnetwork/agent.go#L529-L534
[2]: https://github.com/moby/moby/blob/5b53ddfcdd1c0721f875d3e237078ab7de891a57/libnetwork/agent.go#L688-L696
[3]: https://github.com/moby/libnetwork/commit/f2307265c77fd6db5d92310857c3ffad1184ad24
[4]: https://github.com/moby/libnetwork/commit/6426d1e66f33c0b0c8bb135b7ee547447f54d043
[5]: https://github.com/moby/libnetwork/commit/8dcf9960aa81767376e580f20d9ecea3e68f62ad
[6]: https://github.com/moby/moby/issues/28712
[7]: https://github.com/moby/moby/blob/75fd88ba8938f120ce1f2a21b19cf13c12071b16/vendor/github.com/docker/libnetwork/agent.go#L452

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>