libnetwork: Controller.handleNodeTableEvent: skip networkdb.UpdateEvent
Controller.handleNodeTableEvent was logging an error if it would handle
a `networkdb.UpdateEvent`, however, the function would still continue
processing the event, and attempt to unmarshal an empty `[]byte` to a
`networkdb.NodeAddr` (resulting in an `unexpected end of JSON input` error).

If it _would_ proceed (i.e., if the unmarshal would not error), it would;

- call `Controller.processNodeDiscovery` (with an empty list of IPs)
- which would iterate over all registered drivers, calling `Controller.pushNodeDiscovery`
  for drivers that implement `discoverapi.Discover`.
- ultimately ending up in `Controller.pushNodeDiscovery` returning early
  because no IPs were provided.

Based on the above, it looks like returning early here is the right thing
to do.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>