libnetwork: Controller.pushNodeDiscovery: take single IP, ignore nil
The signature to take multiple addresses originates in [libnetwork/188][1]
("Libnetwork Host Discovery using Swarm Discovery pkg), which initially
had a `diff` function as part of the `hostdiscovery` package, which would
produce a list of changes in IP-addresses (added, removed).

Discovery with external k/v stores was removed, and the `hostdiscovery`
package was removed in 00f9b23c3ae1614b648e9da1c6423c381f58153e, and
current code always pass a single IP-address (if any).

The Controller.pushNodeDiscovery method contained a check whether the
passed list of node-addresses was nil, however, this was never the
case because `Controller.handleNodeTableEvent` would unconditionally
[initialize a slice][2] (even if the event had a `nil` IP-address).

In case of a `nil` IP-address, this address would be;

- compared with the agent's advertiseAddr to determine if this is `self`
- encoded as a string, and sent as `NodeDiscoveryData.Address`. This string
  would be a literal `"<nil>"`.

While I don't know if there's other implementations that treat this information
different, looking at the [overlay implementation of `driver.DiscoverNew`][3],
it would check the field to be an empty string, and otherwise produce an
error (which wouldn't apply if it contained `"<nil>"`), and set the
driver's advertiseAddress to this value, effectively un-setting it
(`net.ParseIP()` silently ignores invalid values), or producing an error
(`netip.ParseAddr()` will error).

For completeness: `driver.DiscoverDelete` has no implementation in the
overlay driver, but both `driver.DiscoverNew` and `driver.DiscoverDelete`
are both implemented by the `remote` driver, which sends them to the
remote "as-is", so it will depend on the receiving end to handle.

So based on the overlay implementation, my assumption is that an empty
node-address should be considered _invalid_ (both for `DiscoverDelete`
and `DiscoverNew`), so let's return early if we encounter one, but log
a warning to discover cases where this happened.

[1]: https://github.com/moby/libnetwork/commit/ac4d7b61367e34ac341c2171d70d8b6a0a0fbecf
[2]: https://github.com/moby/moby/blob/5b53ddfcdd1c0721f875d3e237078ab7de891a57/libnetwork/agent.go#L905
[3]: https://github.com/moby/moby/blob/5b53ddfcdd1c0721f875d3e237078ab7de891a57/libnetwork/drivers/overlay/overlay.go#L92-L97

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>