daemon: convert short network ID into full name
The daemon already expect in many places that endpoints in
`container.NetworkSettings.Networks` are keyed by network's full name,
once the endpoint has been effectively created.

Also, it already handles full ID -> full name conversion but it doesn't
properly handle the case where user-specified settings where stored with
a partial network ID key.

Because of this missing case, if a container is created with a reference
to a partial network ID, two entries for the same network are inserted
in `container.NetworkSettings.Networks`.

Beside bad UX (ie. duplicated entry when inspecting the container),
there's another concrete issue: when such container is disconnected from
a network and then restarted, the container will be unintendedly
reconnected to that network.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>